+++ CHANGELOG.md
@@ -5,10 +5,12 @@ CHANGELOG
 # Next [1.0.6]
 
 # Env Vars
+- `STEAM_API_KEY`: a steam API key, see https://steamcommunity.com/dev
 
 # Bugfixes
 
 # Features
+- /register can now convert between steamid32, steamid64, a full profile URL or the vanity part of the profile URL
 
 # Commands
-
+- /profile to get a link to your or someone else's steam profile